backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor>=3.5,<4
pydantic==2.7.1
pydantic-settings==2.2.1
bcrypt>=4.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-multipart==0.0.9
email-validator==2.1.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
mongomock-motor==0.0.21
